Explore: keep the URL in step with the pano, POV and immersive state (#5480)
The address bar follows the labeler, so it is always a shareable link to
where they are and what they see: on every pano change, and on POV changes
at most twice a second, ExploreUrlSync rewrites the query with panoId, lat,
lng, heading, pitch, zoom and, in immersive mode, immersive=1 -- the params
ExploreController.explore already reads. replaceState, never pushState, so
a pan is not a Back entry; writes are throttled and guarded because Safari
throws past 100 history writes in 30 s.

The URL names a place, not a session: the mission's own params (routeId,
resumeRoute, regionId, streetEdgeId) and the drop-in greeting's placeName
are dropped once the page is up, so a recipient lands in free exploration
at that spot, never in the sharer's mission or route. The tutorial keeps
its old one-shot cleanup.

ImmersiveMode reads immersive=1 from the URL (and keeps it for the sitting,
like a toggle would) and takes an onChange hook so a toggle updates the URL
at once; ImmersiveMode_Restored gains a source note.

zoom on /explore and /audit widens from Option[Int] to Option[Double]: the
scroll-wheel zoom is continuous, and an Int binder would 400 the live URL
on reload.

+++ app/controllers/ExploreController.scala
@@ -58,7 +58,7 @@ class ExploreController @Inject() (
       placeName: Option[String],
       heading: Option[Double],
       pitch: Option[Double],
-      zoom: Option[Int]
+      zoom: Option[Double]
   ) = cc.securityService.SecuredAction { implicit request =>
     val user: SidewalkUserWithRole = request.identity
